% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.r
\docType{data}
\name{cddd}
\alias{cddd}
\title{Change Directory Dot Dot.}
\format{\preformatted{ chr "/home/xavier/code/R/lkdjiinr"
}}
\usage{
cddd
}
\description{
Move up one directory level. Same as UNIX `cd ..`.
}
\examples{
cddd()
}
\keyword{datasets}

