% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/string.r
\name{\%+\%}
\alias{\%+\%}
\title{String Concatenation Operator.}
\usage{
... \%+\% NA
}
\value{
A string.
}
\description{
The world is in need of an operator to concatenate strings. This will
be %+%.
}
\examples{
"foo" \%+\% "bar"
#=> "foobar"

c("a", "a") \%+\% c("b", "c")
#=> "ab" "ac"
}

