% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/displaying.r
\name{println}
\alias{println}
\title{Print with a newline.}
\usage{
println(x)
}
\arguments{
\item{x}{The object to display.}
}
\description{
Just like \code{cat} with an automatic newline.
}
\examples{
println('foo')
}

