% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/string.r
\name{tr}
\alias{tr}
\title{Alias to chartr.}
\usage{
tr(x, old, new)
}
\arguments{
\item{x}{A vector of strings to transform.}

\item{old}{Old characters.}

\item{new}{New characters.}
}
\value{
A vector of strings.
}
\description{
Like \code{chartr} but with a different order of arguments.
}
\examples{
tr("foo", "o", "u")
#=> "fuu"
}
\seealso{
\code{\link{chartr}}
}

