% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/displaying.r
\name{tw}
\alias{tw}
\title{Terminal Width.}
\usage{
tw()
}
\description{
Update the width of the terminal. R use 80 columns by default, no
matter what the width of your terminal is. Calling \code{tw} set the
R's width according to your current terminal's width. If you resize
your terminal, you'll have to recall \code{tw} again.
}
\examples{
tw()
}

