% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.r
\name{na_count}
\alias{na_count}
\title{Count number of NAs in a vector.}
\usage{
na_count(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The number of NAs in the vector x.
}
\description{
Count number of NAs in a vector.
}
\examples{
x <- c(NA, 'a', 'b', NA, 'c')
na_count(x)
# 2
}
\seealso{
\code{\link{nas}}
}

