% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.r
\name{na_percent}
\alias{na_percent}
\title{Percentage of NAs in a vector.}
\usage{
na_percent(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A number between 0 (no NAs) and 100 (all values are NAs).
}
\description{
What percentage of NA values are in a vector?
}
\examples{
x <- c(NA, 'a', 'b', NA, 'c')
na_percent(x)
#=> 40
}
\seealso{
\code{\link{na_prop}}
}

