% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.r
\name{na_prop}
\alias{na_prop}
\title{Proportion of NAs in a vector.}
\usage{
na_prop(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A number between 0 (no NAs) and 1 (all values are NAs).
}
\description{
What proportion of NA values are in a vector?
}
\examples{
x <- c(NA, 'a', 'b', NA, 'c')
na_prop(x)
#=> 0.4
}
\seealso{
\code{\link{na_percent}}
}

